requirements.txt
numpy
opencv-python
Pillow
tensorboard
gin-config
rawpy
mediapy
matplotlib
absl-py
# would be integrated into pytorch: https://github.com/pytorch/pytorch/pull/93139
optree==0.9.0
torch==2.0.0